Added Redirect and Null Checks

+++ classes/learner_task_submission.php
@@ -549,16 +549,18 @@ class learner_task_submission extends learner_task_submission_base implements te
         ];
     }
 
-    public function delete() {
-        foreach($this->learner_attempts as $learner_task_attempt) {
+    public function delete()
+    {
+        foreach ($this->learner_attempts as $learner_task_attempt) {
             $learner_task_attempt->delete();
         }
 
         foreach ($this->observer_assignments as $observer_assignment_task) {
             $observer_assignment_task->delete();
         }
-        $this->assessor_task_submission->delete();
-
+        if (!is_null($this->assessor_task_submission)) {
+            $this->assessor_task_submission->delete();
+        }
         parent::delete();
     }
 }
